requirements.txt
selenium
webdriver-manager
beautifulsoup4
pandas
gspread
google-api-python-client
google-auth
pytesseract
Pillow
requests
python-dotenv
pyperclip
